@@ -1,13 +1,12 @@
-﻿using Amazon.SimpleSystemsManagement;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using Amazon.SimpleSystemsManagement;
 using Amazon.SimpleSystemsManagement.Model;
 using Microsoft.Extensions.Configuration;
-using System.Linq;
-using System;
-using System.Collections.Generic;
 
 namespace ParameterStoreConfigurationProvider
 {
-
     public class ParameterStoreConfigurationProvider : ConfigurationProvider
     {
         private readonly ParameterStoreConfigurationSource configurationSource;
@@ -21,18 +20,17 @@ namespace ParameterStoreConfigurationProvider
         {
             IEnumerable<GetParametersResponse> responses;
 
-            if (this.configurationSource.UseDefaultCredentials)
+            if (configurationSource.UseDefaultCredentials)
             {
-                using (var client = new AmazonSimpleSystemsManagementClient(Amazon.RegionEndpoint.GetBySystemName(this.configurationSource.Region)))
+                using (var client = new AmazonSimpleSystemsManagementClient(Amazon.RegionEndpoint.GetBySystemName(configurationSource.Region)))
                 {
                     responses = MappingClientResponseToData(client);
                 }
             }
             else
             {
-                using (var client = new AmazonSimpleSystemsManagementClient(this.configurationSource.AwsCredential, Amazon.RegionEndpoint.GetBySystemName(this.configurationSource.Region)))
+                using (var client = new AmazonSimpleSystemsManagementClient(configurationSource.AwsCredential, Amazon.RegionEndpoint.GetBySystemName(configurationSource.Region)))
                 {
-
                     responses = MappingClientResponseToData(client);
                 }
             }
@@ -40,9 +38,9 @@ namespace ParameterStoreConfigurationProvider
             MapResults(responses);
         }
 
-        private IEnumerable<GetParametersResponse> MappingClientResponseToData(AmazonSimpleSystemsManagementClient client)
+        private IEnumerable<GetParametersResponse> MappingClientResponseToData(IAmazonSimpleSystemsManagement client)
         {
-            IEnumerable<GetParametersRequest> requests = PrepareRequests();
+            var requests = PrepareRequests();
             IList<GetParametersResponse> responses = new List<GetParametersResponse>();
 
             foreach (var request in requests)
@@ -58,8 +56,9 @@ namespace ParameterStoreConfigurationProvider
         private void CheckParametersValidity(GetParametersResponse response)
         {
             var requiredInvalidParameters = response.InvalidParameters.Where(item =>
-                this.configurationSource.ParameterMapping.First(pm =>
+                configurationSource.ParameterMapping.First(pm =>
                     pm.AwsName == item).Optional == false).ToList();
+
             if (requiredInvalidParameters.Count > 0)
             {
                 var wrongParams = "";
@@ -72,9 +71,9 @@ namespace ParameterStoreConfigurationProvider
             }
         }
 
-        internal IEnumerable<GetParametersRequest> PrepareRequests()
+        private IEnumerable<GetParametersRequest> PrepareRequests()
         {
-            var names = this.configurationSource.ParameterMapping.Select(x => x.AwsName).ToList();
+            var names = configurationSource.ParameterMapping.Select(x => x.AwsName).ToList();
             const int groupSize = 10;
 
             var requests = names
@@ -83,30 +82,30 @@ namespace ParameterStoreConfigurationProvider
                 .Select(g => new GetParametersRequest
                 {
                     Names = g.ToList(),
-                    WithDecryption = this.configurationSource.WithDecryption
+                    WithDecryption = configurationSource.WithDecryption
                 });
 
             return requests;
         }
 
-        internal void MapResults(IEnumerable<GetParametersResponse> responses)
+        private void MapResults(IEnumerable<GetParametersResponse> responses)
         {
-            this.Data = new Dictionary<string, string>();
+            Data = new Dictionary<string, string>();
 
             foreach (var response in responses)
             {
                 foreach (var parameter in response.Parameters)
                 {
                     var parameterMapping =
-                        this.configurationSource.ParameterMapping.First(pm => pm.AwsName == parameter.Name);
-                    this.Data[parameterMapping.SettingName] = parameter.Value;
+                        configurationSource.ParameterMapping.First(pm => pm.AwsName == parameter.Name);
+                    Data[parameterMapping.SettingName] = parameter.Value;
                 }
 
                 foreach (var parameter in response.InvalidParameters)
                 {
                     var parameterMapping =
-                        this.configurationSource.ParameterMapping.First(pm => pm.AwsName == parameter);
-                    this.Data[parameterMapping.SettingName] = parameterMapping.Default;
+                        configurationSource.ParameterMapping.First(pm => pm.AwsName == parameter);
+                    Data[parameterMapping.SettingName] = parameterMapping.Default;
                 }
             }
         }
