@@ -1,14 +1,12 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 
 namespace example_api.Controllers
 {
-    [Route("api/[controller]")]
-    public class ConfigurationController : Controller
+    [ApiController]
+    [Route("[controller]")]
+    public class ConfigurationController : ControllerBase
     {
         private readonly IConfiguration configuration;
 
@@ -17,7 +15,7 @@ namespace example_api.Controllers
             this.configuration = configuration;
         }
 
-        // GET api/values/5
+        // GET Configuration/key
         [HttpGet("{key}")]
         public string Get(string key)
         {
