@@ -1,12 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Hosting;
 using ParameterStoreConfigurationProvider;
 
 namespace example_api
@@ -15,14 +11,13 @@ namespace example_api
     {
         public static void Main(string[] args)
         {
-            BuildWebHost(args).Run();
+            BuildHost(args).Run();
         }
 
-        public static IWebHost BuildWebHost(string[] args)
+        public static IHost BuildHost(string[] args)
         {
-            
-            return WebHost.CreateDefaultBuilder(args)
-                    .ConfigureAppConfiguration((hostContext, config)=>
+            return Host.CreateDefaultBuilder(args)
+                    .ConfigureAppConfiguration((hostContext, config) =>
                     {
                         config.SetBasePath(Directory.GetCurrentDirectory())
                         .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
@@ -36,7 +31,7 @@ namespace example_api
                             };
                             parameterStoreConfig.Region = "eu-west-1";
                             parameterStoreConfig.UseDefaultCredentials = true;
-                       //   parameterStoreConfig.AwsCredential = new Amazon.Runtime.StoredProfileAWSCredentials();
+                            //   parameterStoreConfig.AwsCredential = new Amazon.Runtime.StoredProfileAWSCredentials();
                         })
                         .AddParameterStoreConfig(parameterStoreConfig =>
                         {
@@ -47,10 +42,13 @@ namespace example_api
                             parameterStoreConfig.WithDecryption = true;
                             parameterStoreConfig.Region = "eu-west-1";
                             parameterStoreConfig.UseDefaultCredentials = true;
-                        //    parameterStoreConfig.AwsCredential = new Amazon.Runtime.StoredProfileAWSCredentials();
+                            //    parameterStoreConfig.AwsCredential = new Amazon.Runtime.StoredProfileAWSCredentials();
                         });
                     })
-                    .UseStartup<Startup>()
+                    .ConfigureWebHostDefaults(webBuilder =>
+                    {
+                        webBuilder.UseStartup<Startup>();
+                    })
                     .Build();
         }
     }
